BC-3961 - rename module
