+++ apps/server/src/modules/files-storage/helper/test-helper.ts
@@ -1,4 +1,4 @@
-import { GetFile } from '@shared/infra/s3-file-storage';
+import { GetFile } from '@shared/infra/s3-client';
 import { Readable } from 'stream';
 import { GetFileResponse } from '../interface';
