+++ apps/server/src/modules/files-storage/service/files-storage-download.service.spec.ts
@@ -4,7 +4,7 @@ import { NotAcceptableException, NotFoundException } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { Test, TestingModule } from '@nestjs/testing';
 import { AntivirusService } from '@shared/infra/antivirus/antivirus.service';
-import { GetFile, S3ClientAdapter } from '@shared/infra/s3-file-storage';
+import { GetFile, S3ClientAdapter } from '@shared/infra/s3-client';
 import { fileRecordFactory, setupEntities } from '@shared/testing';
 import { LegacyLogger } from '@src/core/logger';
 import { FileRecordParams } from '../controller/dto';
