+++ apps/server/src/shared/infra/s3-client/index.ts
@@ -1,3 +1,3 @@
 export * from './interface';
 export * from './s3-client.adapter';
-export * from './s3-file-storage.module';
+export * from './s3-client.module';
