+++ apps/server/src/shared/infra/s3-client/interface/index.ts
