+++ apps/server/src/shared/infra/s3-client/s3-client.adapter.ts
